% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.schedule}
\alias{print.schedule}
\title{Print Randomization Schedule, Envelopes and Inserts}
\usage{
\method{print}{schedule}(
  x,
  path = tempdir(),
  title = "Study Title",
  scientist = "Principal.Investigator@jefferson.edu",
  statistician = "Tingting.Zhan@jefferson.edu",
  ...
)
}
\arguments{
\item{x}{a \link{schedule} object}

\item{path}{\link[base]{character} scalar}

\item{title}{\link[base]{character} scalar, name of study}

\item{scientist}{\link[base]{character} scalar, (email of) the principal investigator}

\item{statistician}{\link[base]{character} scalar, (email of) the statistician}

\item{...}{..}
}
\value{
Function \link{print.schedule} returns ..
}
\description{
Create randomization envelopes and inserts using \link[ggplot2]{ggplot}.
}
\examples{
# see ?schedule

}
