% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randEnvelope.R
\name{randEnvelope}
\alias{randEnvelope}
\title{Randomization Envelopes and Inserts}
\usage{
randEnvelope(x, path = tempdir(), ...)
}
\arguments{
\item{x}{a \link{randSchedule} object}

\item{path}{\link[base]{character} scalar}

\item{...}{..}
}
\value{
Function \link{randEnvelope} returns ..
}
\description{
Create randomization envelopes and inserts using \link[ggplot2]{ggplot}.
}
\examples{
# see ?randSchedule

}
