% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSchedule.R
\name{randSchedule}
\alias{randSchedule}
\title{(Stratified) Permuted Block Randomization Schedules}
\usage{
randSchedule(
  treatment,
  tx_ratio = rep(1L, times = ntx),
  strata,
  strata_sep = " ֍ ",
  n = stop("number of patients (per stratum) is required"),
  mblock = 2:3,
  title = "Study Title",
  scientist = "Principal.Investigator@jefferson.edu",
  statistician = "Tingting.Zhan@jefferson.edu",
  ...
)
}
\arguments{
\item{treatment}{\link[base]{character} \link[base]{vector}, name of treatments,
such as \code{'tx1'}, \code{'tx2'}, \code{'control'}, etc.}

\item{tx_ratio}{\link[base]{integer} \link[base]{vector},
planned allocation ratios of the \code{treatment} arms.
End user should make sure the greatest common divisor of \code{tx_ratio} is 1.
Default to balanced design \code{1:1: ... :1}.}

\item{strata}{(optional) named \link[base]{list} of \link[base]{character} \link[base]{vector}s to denote the strata.
Default to no stratification (missing argument).}

\item{strata_sep}{\link[base]{character} scalar, symbol to separate multiple strata names}

\item{n}{\link[base]{integer} scalar, smallest sample size to be guaranteed.
In stratified randomization, this is the size \emph{per stratum}.}

\item{mblock}{\link[base]{integer} \link[base]{vector} of \link[base]{length} \eqn{>1},
\emph{multipliers} to determine the permuted block sizes.
Default \code{2:3}, i.e., block size multipliers of 2 and 3 are permuted.}

\item{title}{\link[base]{character} scalar, name of study}

\item{scientist}{\link[base]{character} scalar, (email of) the principal investigator}

\item{statistician}{\link[base]{character} scalar, (email of) the statistician}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{randSchedule} returns a \link[base]{data.frame}
}
\description{
..
}
\details{
\describe{

\item{Simple randomization}{Based on a single sequence of random assignments.}

\item{Block randomization}{Randomize subjects into groups of equal sample sizes (balance in sample size across groups).
Blocks are small and balanced (or with pre-determined group-ratio) with predetermined group assignments.
\enumerate{
\item{Block size is pre-determined and is a small multiple of the number of groups (e.g. *2, *3, *4, etc if balanced with group);}
\item{All possible combinations of groups within the block are calculated;}
\item{Blocks are then randomly chosen.}
}}

\item{Stratified randomization}{Control and balance the influence of covariates (i.e. pre-determined subjects’ baseline characteristics) among groups
\enumerate{
\item{Generate a separate block for each combination of covariates (with subjects assigned to the appropriate block of covariates);}
\item{Simple randomization within each block to assign subjects into groups.}
}}

}
}
\examples{
tx = c('treatment', 'control')
txr = c(1L, 2L)
strt = list(cohort = c('non_Oral', 'Oral'), institute = c('Ohio', 'Jeff'))

r1 = randSchedule(tx, txr, n = 40L)
r2 = randSchedule(tx, txr, strata = strt, n = 30L)

if (TRUE) {
file.remove(randEnvelope(r1))
file.remove(randEnvelope(r2[1:10,]))
}

}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3136079/}
}
