% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{rpermblock}
\alias{rpermblock}
\title{Random Generation based on \linkS4class{permblock}}
\usage{
rpermblock(x)
}
\arguments{
\item{x}{a \linkS4class{permblock}}
}
\value{
The function \code{\link[=rpermblock]{rpermblock()}} returns a \link[base]{character}
\link[base]{vector} with \link[base]{length} no less than \verb{@n}.
}
\description{
Random Generation based on \linkS4class{permblock}
}
\details{
\enumerate{
\item \link[base]{sample}, \emph{with} replacement,  a block-size multiplier.
\item \link[base]{sample}, \emph{without} replacement, in the selected block.
}
}
\note{
The function \link[base]{sample} is not an \code{S3} generic function.

The function \code{\link[=rpermblock]{rpermblock()}} is named in the fashion of the function \link[stats]{rnorm}.
}
