% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_block.R
\name{sample_block}
\alias{sample_block}
\title{Sample from Blocks}
\usage{
sample_block(x, n)
}
\arguments{
\item{x}{\link[base]{list} of depth-2, randomized blocks}

\item{n}{positive \link[base]{integer} scalar, sample size}
}
\description{
Sample from Blocks
}
\examples{
pb = permblock(arm = c('intervention', 'control'), ratio = 1:2, n = 20L)
pb |> get_block() |> sample_block(n = 30L) # base::set.seed if needed
}
\keyword{internal}
