% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.permblock}
\alias{schedule.stratified}
\title{schedule}
\usage{
schedule(x)

\method{schedule}{permblock}(x)

\method{schedule}{stratified}(x)
}
\arguments{
\item{x}{..}
}
\value{
All \code{S3} methods of the generic function \code{\link[=schedule]{schedule()}}
return a \link{schedule} object.
}
\description{
schedule
}
