% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permblock.R
\name{show,permblock-method}
\alias{show,permblock-method}
\title{Show \linkS4class{permblock}}
\usage{
\S4method{show}{permblock}(object)
}
\arguments{
\item{object}{\linkS4class{permblock}}
}
\description{
Show \linkS4class{permblock}
}
