% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{show,stratified-method}
\alias{show,stratified-method}
\title{Show \linkS4class{stratified}}
\usage{
\S4method{show}{stratified}(object)
}
\arguments{
\item{object}{\linkS4class{stratified}}
}
\description{
Show \linkS4class{stratified}
}
