% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{show,stratified_permblock-method}
\alias{show,stratified_permblock-method}
\title{Show \linkS4class{stratified_permblock}}
\usage{
\S4method{show}{stratified_permblock}(object)
}
\arguments{
\item{object}{\linkS4class{stratified_permblock}}
}
\description{
Show \linkS4class{stratified_permblock}
}
