% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\docType{class}
\name{stratified-class}
\alias{stratified-class}
\title{Stratified Permuted Block Randomization}
\description{
Stratified permuted block randomization.
}
\section{Slots}{

\describe{
\item{\code{strata}}{\link[base]{data.frame}}

\item{\code{label}}{\link[base]{character} \link[base]{vector}, strata labels}
}}

