% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\alias{stratify.permblock}
\title{stratify}
\usage{
stratify(x, ..., sep)

\method{stratify}{permblock}(x, ..., sep = " ֍ ")
}
\arguments{
\item{x}{a \linkS4class{permblock}}

\item{...}{one or more \link[base]{character} \link[base]{vector}s
to be \link[base]{expand.grid} into the slot \verb{@strata}}

\item{sep}{\link[base]{character} scalar, symbol to separate multiple strata names}
}
\value{
All methods of the \code{S3} generic function \code{\link[=stratify]{stratify()}}
return \linkS4class{stratified} \linkS4class{permblock}.
}
\description{
stratify
}
