% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permblock.R
\docType{class}
\name{permblock}
\alias{permblock}
\alias{permblock-class}
\title{Permuted Block Randomization Schedule}
\usage{
permblock(...)
}
\arguments{
\item{...}{slots of S4 object \linkS4class{permblock}}
}
\description{
Permuted block randomization schedule, non-stratified.
}
\section{Slots}{

\describe{
\item{\code{arm}}{\link[base]{character} \link[base]{vector}, name of arms,
such as \code{'tx1'}, \code{'tx2'}, \code{'control'}, etc.}

\item{\code{ratio}}{\link[base]{integer} \link[base]{vector},
planned allocation ratios of arms.
End user should make sure the greatest common divisor of \code{ratio} is 1.
Default to balanced design \code{1:1: ... :1}.}

\item{\code{n}}{\link[base]{integer} scalar, smallest sample size to be guaranteed.}

\item{\code{multiplier}}{\link[base]{integer} \link[base]{vector}
to determine the permuted block sizes.
Default \code{2:3}, i.e., block size multipliers of 2 and 3 are permuted.}

\item{\code{nm_sequence}}{\link[base]{character} scalar, default \code{'Sequence'}}

\item{\code{nm_arm}}{\link[base]{character} scalar, default \code{'Assignment'}}
}}

\examples{
permblock(arm = c('intervention', 'control'), ratio = 1:2, n = 20L)
}
