% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.schedule}
\alias{print.schedule}
\title{Print Randomization Schedule, Envelopes and Inserts}
\usage{
\method{print}{schedule}(
  x,
  path = tempdir(),
  which = c("schedule", "envelope", "insert", "all"),
  mc.cores = getOption("cores"),
  title = "Study Title",
  scientist = "Principal.Investigator@jefferson.edu",
  statistician = person(given = "Tingting", family = "Zhan", email =
    "Tingting.Zhan@jefferson.edu"),
  ...
)
}
\arguments{
\item{x}{a \link{schedule} object}

\item{path}{\link[base]{character} scalar}

\item{which}{\link[base]{character} scalar, either one of
\describe{
\item{\code{'schedule'}}{(default) the randomization schedule in a \code{.csv} file,
to be uploaded to a REDCap data base.
This procedure is fast.}
\item{\code{'envelope'} or \code{'insert'}}{the randomization envelopes and inserts
in \code{.pdf} files, printed using the package \CRANpkg{ggplot2}.
This procecure could be slow for large sample size.}
\item{\code{'all'}}{all options above}
}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is the return of function \link[parallel]{detectCores}.}

\item{title}{\link[base]{character} scalar, name of study}

\item{scientist}{\link[base]{character} scalar, (email of) the principal investigator}

\item{statistician}{(an R object convertible to) \link[utils]{person} object}

\item{...}{..}
}
\value{
The \code{S3} method \code{\link[=print.schedule]{print.schedule()}} returns ..
}
\description{
Create randomization envelopes and inserts
using the package \CRANpkg{ggplot2}.
}
