% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.schedule}
\alias{print.schedule}
\title{Print Randomization Schedule, Envelopes and Inserts}
\usage{
\method{print}{schedule}(
  x,
  path = tempdir(),
  which = c("schedule", "envelope", "insert", "all"),
  mc.cores = getOption("cores"),
  title = "Study Title",
  scientist = "Principal.Investigator@jefferson.edu",
  statistician = "Tingting.Zhan@jefferson.edu",
  ...
)
}
\arguments{
\item{x}{a \link{schedule} object}

\item{path}{\link[base]{character} scalar}

\item{which}{\link[base]{character} scalar, either
\code{'schedule'} (default, in a \code{.csv} file, fast),
\code{'envelope'},
\code{'insert'} (in \code{.pdf} files via \link[ggplot2]{ggplot}, could be slow for large sample size)
and \code{'all'}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is the return of function \link[parallel]{detectCores}.}

\item{title}{\link[base]{character} scalar, name of study}

\item{scientist}{\link[base]{character} scalar, (email of) the principal investigator}

\item{statistician}{\link[base]{character} scalar, (email of) the statistician}

\item{...}{..}
}
\value{
Function \code{\link[=print.schedule]{print.schedule()}} returns ..
}
\description{
Create randomization envelopes and inserts using \link[ggplot2]{ggplot}.
}
