% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{randEnvelope-package}
\alias{randEnvelope}
\alias{randEnvelope-package}
\title{randEnvelope}
\description{
..
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
