% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{rpermblock}
\alias{rpermblock}
\title{Random Generation based on \linkS4class{permblock}}
\usage{
rpermblock(x)
}
\arguments{
\item{x}{a \linkS4class{permblock}}
}
\value{
Function \code{\link[=rpermblock]{rpermblock()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Random Generation based on \linkS4class{permblock}
}
\details{
First select a block-size multiplier.

Then do a \link[base]{sample}ing without replacement in the selected block.
}
