% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.permblock}
\alias{schedule.stratified_permblock}
\title{schedule}
\usage{
schedule(x, ...)

\method{schedule}{permblock}(x, ...)

\method{schedule}{stratified_permblock}(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{additional information to be appended to the randomization schedule.
Use with extreme caution!}
}
\value{
Functions \link{schedule.permblock} and \link{schedule.stratified_permblock}
both returns a \link{schedule} object
}
\description{
schedule
}
\examples{
(pb = permblock(arm = c('intervention', 'control'), ratio = 1:2, n = 50L))
set.seed(1251); pb |> schedule()
set.seed(1251); pb |> schedule(study.name = 'CDC')
(spb = pb |> 
  stratify(cohort = c('young', 'old'), state = c('PA', 'NJ')))
set.seed(1325); spb |> schedule()
}
