% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.permblock}
\alias{schedule.stratified_permblock}
\title{schedule}
\usage{
schedule(x, ...)

\method{schedule}{permblock}(x, .blocks = get_block(x), ...)

\method{schedule}{stratified_permblock}(x, .blocks = get_block(x), ...)
}
\arguments{
\item{x}{..}

\item{...}{..}

\item{.blocks}{..}
}
\description{
schedule
}
\examples{
pb = permblock(arm = c('intervention', 'control'), ratio = 1:2, n = 20L)
set.seed(1251); r1 = pb |> schedule()
r1
set.seed(1325); r2 = pb |> 
  stratify(cohort = c('young', 'old'), state = c('PA', 'NJ')) |> 
  schedule()
r2
}
