% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.permblock}
\alias{schedule.stratified}
\title{Create an Randomization Schedule}
\usage{
schedule(x)

\method{schedule}{permblock}(x)

\method{schedule}{stratified}(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\value{
All \code{S3} methods of the generic function \code{\link[=schedule]{schedule()}}
return an R object of \code{'schedule'},
which inherits from the class \link[base]{data.frame}.
}
\description{
Create an Randomization Schedule
}
