% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.permblock}
\alias{schedule.stratified_permblock}
\title{schedule}
\usage{
schedule(x, ...)

\method{schedule}{permblock}(x, .blocks = getblocks(x), ...)

\method{schedule}{stratified_permblock}(x, .blocks = getblocks(x), ...)
}
\arguments{
\item{x}{..}

\item{...}{..}

\item{.blocks}{..}
}
\description{
schedule
}
\examples{
pb = permblock(treatment = c('treatment', 'control'), ratio = 1:2, n = 20L)
set.seed(1251); r1 = pb |> schedule()
randEnvelope(r1)
set.seed(1325); r2 = pb |> 
  stratify(cohort = c('non_Oral', 'Oral'), inst = c('Ohio', 'Jeff')) |> 
  schedule()
randEnvelope(r2)
}
