% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.permblock}
\alias{schedule.stratified_permblock}
\title{schedule}
\usage{
schedule(x, ...)

\method{schedule}{permblock}(x, ...)

\method{schedule}{stratified_permblock}(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{additional information to be appended to the randomization schedule.
Use with extreme caution!}
}
\value{
Functions \link{schedule.permblock} and \link{schedule.stratified_permblock}
both returns a \link{schedule} object
}
\description{
schedule
}
