% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\docType{class}
\name{stratified_permblock-class}
\alias{stratified_permblock-class}
\title{stratified_permblock}
\description{
stratified_permblock
}
\section{Slots}{

\describe{
\item{\code{strata}}{\link[base]{data.frame}, e.g., \link[base]{expand.grid} from a \link[base]{list} of \link[base]{character} \link[base]{vector}s.}

\item{\code{label}}{\link[base]{character} \link[base]{vector}, strata labels}
}}

