% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\alias{stratify.permblock}
\title{stratify}
\usage{
stratify(x, ..., sep)

\method{stratify}{permblock}(x, ..., sep = " Ö ")
}
\arguments{
\item{x}{..}

\item{...}{..}

\item{sep}{\link[base]{character} scalar, symbol to separate multiple strata names}
}
\value{
Function \link{stratify.permblock} returns a \linkS4class{stratified_permblock} object.
}
\description{
stratify
}
\examples{
permblock(arm = c('intervention', 'control'), ratio = 1:2, n = 20L) |>
 stratify(cohort = c('young', 'old'), inst = c('PA', 'NJ'))
}
