% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\alias{stratify.permblock}
\title{stratify}
\usage{
stratify(x, ..., sep)

\method{stratify}{permblock}(x, ..., sep = " ֍ ")
}
\arguments{
\item{x}{..}

\item{...}{..}

\item{sep}{\link[base]{character} scalar, symbol to separate multiple strata names}
}
\value{
Function \code{\link[=stratify.permblock]{stratify.permblock()}} returns a \linkS4class{stratified_permblock} object.
}
\description{
stratify
}
